add reload timer to projectile spawner
This is done so that player cannot spawn projectiles
infinitely fast.

It is required to include a Timer node as a child of
the spawning node. For example, any scene that the player
scene is added to, a Timer must be added as a child to
the player character node.
This might be due to the fact that there is no spawner scene
which exists only as a script... I am not sure, but it
works this way.